requirements.txt
flask==2.2.3
neo4j==5.8.1
python-dotenv==1.0.0
flask-cors==3.0.10
gunicorn==20.1.0